+++ pkg/helpers/browser.go
@@ -1,0 +1,25 @@
+// Copyright Contributors to the Open Cluster Management project
+
+package helpers
+
+import (
+	"fmt"
+	"os/exec"
+	"runtime"
+)
+
+func Openbrowser(url string) error {
+	var err error
+
+	switch runtime.GOOS {
+	case "linux":
+		err = exec.Command("xdg-open", url).Start()
+	case "windows":
+		err = exec.Command("rundll32", "url.dll,FileProtocolHandler", url).Start()
+	case "darwin":
+		err = exec.Command("open", url).Start()
+	default:
+		err = fmt.Errorf("unsupported platform")
+	}
+	return err
+}