+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 # Release Content
 
 - [Add examples in the clusterpool.md or in the help of the command for cm create cph](https://github.com/open-cluster-management/cm-cli/issues/59)
+- Add check on RHACM version when running `cm attach clusterclaim`
 
 ### Additions
+
 ### Breacking changes
 ### Bug fixes
 - [Columns are not well aligned when running cm get cc and cp](https://github.com/open-cluster-management/cm-cli/issues/58)